chore: Fix 'npm install' failing in CI and locally
The npm "prepublish" script is nasty in that it is also ran after "npm
install" is ran. And as the last step in that script is "git commit",
and if the working tree is clean, that command will return an exit code.
This in turn will make the CI fail